requirements.txt
praw
pandas
python-dotenv
celery
redis
